Move the MusicChannel enum to the consts module
This allows it to be used by the compiler
